fix(gate): ungate the transcript generator tests so the merge gate runs them
Issue #170: TestTranscriptReadersAgreeSanity and
TestStructuredTranscriptReachesDeeper were doubly invisible to the merge
gate - named in GATE_FUZZ_TEST_SKIP and behind //go:build evenerfuzz, so
the gate never even compiled them. Removing the skip entries alone (PR
#221) was a no-op. Per the RCA spec on #170:

- Split the transcript generator and its two deterministic evidence
  tests out of the tagged fuzz files into untagged
  transcript_structured_gen_test.go, so the default build (and thus the
  gate) compiles and runs them. Only the native Fuzz* entry points stay
  behind the tag; turnsOf/jsonEqual and the reader-agreement helpers
  move along because the untagged tests need them.
- Rename TestStructuredTranscriptReachesDeeper to
  TestTranscriptGenReachesDeeper so the Structured.*Reach skip entry
  keeps excluding the five other still-fuzz-only reachability proofs
  without regex surgery.
- Bound the Monte Carlo at 300 samples under testing.Short() (the
  gate's mode); the full 3000 still runs under make fuzz's registry row
  (renamed in fuzz-targets.txt) and plain go test.
- Drop only TranscriptReadersAgreeSanity from GATE_FUZZ_TEST_SKIP.
- Delete toolNameList: dead in every build config, surfaced by lint
  once the file left the tag.

Gate-config evidence: before, go test -run '^(TestTranscriptReadersAgree
Sanity|TestStructuredTranscriptReachesDeeper)$' ./agent -> no tests to
run; after, both pass untagged (2.97s full, 0.09s under -short) and
under -tags evenerfuzz (make fuzz path). Skip-set delta verified: only
TranscriptReadersAgreeSanity leaves the skip set.
